feat: allow style set by environment variable (#43)
This is the standard environment variable used with other Glamour apps
(such as Github CLI).  This way the user does not have to add extra
arguments (--style=) to all their comand line invocations to use a style
file, and the value of the variable will get passed to the Glamour
backend for interpretation.

Furthermore, it allows sharing settings with other Glamour-using
applications, by using an environment variable.